requirements.txt
pandas
requests
pymongo
psycopg2-binary
python-dotenv
apache-airflow